+++ server/server.js
@@ -9,15 +9,12 @@ const customersRouter = require('./routes/customers');
 const postsRouter = require('./routes/posts');
 const vendorsRouter = require('./routes/vendors');
 
-const validator = require('validator');
 const bodyParser = require('body-parser');
 const bcrypt = require('bcryptjs');
 const passport = require('passport');
 const session = require('express-session');
 const LocalStrategy = require('passport-local').Strategy;
 const redis = require('connect-redis')(session);
-
-const saltRounds = 12;
 
 app.use(bodyParser.json({ extended: true }));
 app.use(bodyParser.urlencoded({ extended: true }));
@@ -66,9 +63,11 @@ passport.use(new LocalStrategy((username, password, done) => {
       }
       else {
         let customerObj = user.serialize();
+        console.log('passport', customerObj, password)
         bcrypt.compare(password, customerObj.password)
           .then((res) => {
-            if (res) { return done(null, customerObj); }
+            console.log(res)
+            if (res === true) { return done(null, customerObj); }
             else {
               return done(null, false, { message: `Bad username/password` })
             }
@@ -76,87 +75,6 @@ passport.use(new LocalStrategy((username, password, done) => {
       }
     });
 }));
-
-app.post('/api/customers', (req, res) => {
-  console.log('server reg', req.body)
-  let { first_name, last_name, username, password, email, state, city, zip_code } = req.body;
-
-  const parseZipcode = parseInt(zip_code);
-
-  if (!validator.isAlpha(first_name)) {
-    return res.status(400).json({ status: Error, message: 'Invalid first name' });
-  } else if (!validator.isAlpha(last_name)) {
-    return res.status(400).json({ status: Error, message: 'Invalid last name' });
-  } else if (!validator.isAlphanumeric(username)) {
-    return res.status(400).json({ status: Error, message: 'Invalid username, letters and numbers only' });
-  } else if (!validator.isAscii(password)) {
-    return res.status(400).json({ status: Error, message: 'Invalid password' });
-  } else if (!validator.isEmail(email)) {
-    return res.status(400).json({ status: Error, message: 'Invalid email' });
-  } else if (!validator.isAlpha(state) && (state.length !== 2)) {
-    return res.status(400).json({ status: Error, message: 'Invalid state' });
-  } else if (!validator.isAlpha(city)) {
-    return res.status(400).json({ status: Error, message: 'Invalid city' });
-  } else if (!validator.isNumeric(zip_code)) {
-    return res.status(400).json({ status: Error, message: 'Invalid zipcode' });
-  }
-  bcrypt.genSalt(saltRounds, function (err, salt) {
-    bcrypt.hash(password, salt, function (err, hash) {
-      {
-        return new Customer({
-
-          first_name,
-          last_name,
-          username,
-          password: hash,
-          email,
-          state,
-          city,
-          zip_code: parseZipcode
-        })
-          .save()
-          .then(customer => {
-            return res.json(customer);
-          })
-          .catch(err => {
-            return res.status(500).json({ message: err.message, code: err.code });
-          });
-      }
-    });
-  })
-})
-
-app.post('/api/vendors/register', (req, res) => {
-  let { first_name, last_name, company_name, email, password, street_address, city, state, zip_code, photo, website, description, phone_number, license_number } = req.body
-  const parseZip = parseInt(zip_code);
-  console.log(parseZip);
-  //const parsePhone = parseInt(phone_number)
-  bcrypt.genSalt(saltRounds, function (err, salt) {
-    bcrypt.hash(password, salt, function (err, hash) {
-      return new Vendor({
-        first_name,
-        last_name,
-        company_name,
-        email,
-        phone_number,
-        password: hash,
-        street_address,
-        city,
-        state,
-        zip_code: parseZip,
-        photo,
-        website,
-        description,
-        license_number
-      })
-        .save()
-        .then(() => {
-          console.log('route');
-        })
-    })
-  })
-})
-
 
 app.post(`/api/vendors/login`, (req, res) => {
   let = { company_name, password, id, first_name, last_name, email, street_address, city, state, zip_code, phone_number, photo, website, description, license_number } = req.body
@@ -175,33 +93,13 @@ app.post(`/api/vendors/login`, (req, res) => {
     })
 })
 
-app.post(`/api/login`, (req, res) => {
-  let = { username, password, id, email, first_name, last_name, city, state, zip_code, } = req.body
-  return new Customer()
-    .where({ username: username, password: password })
-    .fetch({
-      columns: ["username", "password", "id", "email", "first_name", "last_name", "city", "state", "zip_code"]
-    })
-    .then(data => {
-      if (!data) {
-        return res.status(401).json({ message: `Username or password incorrect` })
-      } else {
-        const customer = data.toJSON();
-        return res.send(customer)
-      }
-    })
-    .catch(err => {
-      return res.send('Username or password is incorrect')
-    })
-})
-app.post('/api/login', passport.authenticate('local', {
-  successRedirect: '/home',
-  failureRedirect: '',
-}));
+app.post('/api/customer/login', passport.authenticate('local', {failureRedirect: ''}),
+  function(req, res){
+    console.log(req.body)
+    return res.send(req.body)
+  });
 
-app.get('/smoke', (req, res) => {
-  res.send('smoke test')
-})
+
